% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{CVAggplot}
\alias{CVAggplot}
\title{Plot the CVA biplot}
\usage{
CVAggplot(x, which.var = 1:bp$p)
}
\arguments{
\item{x}{Object from CVA}

\item{which.var}{which variable to display on biplot}
}
\description{
Plot the CVA biplot
}
