% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsvd.R
\name{get.GSVD}
\alias{get.GSVD}
\title{Get GSVD
Get the components of the GSVD decomposition}
\usage{
get.GSVD(A, B)
}
\arguments{
\item{A}{Matrix A}

\item{B}{Matrix B}
}
\value{
Returns components from the GSVD decomposition
}
\description{
Get GSVD
Get the components of the GSVD decomposition
}
