% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\title{Simulated Data}
\format{
A data set with 100 rows and 301 columns
}
\source{
simulated data
}
\description{
\describe{
\item{Class}{Group variable: 0, 1, 2}
\item{X1}{Variable 1}
...
\item{X300}{Variable 300}
}
}
\keyword{datasets}
