% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axes_coordinates.R
\name{axes_coordinates}
\alias{axes_coordinates}
\title{Provide axes coordinates}
\usage{
axes_coordinates(bp, which.var = 1:bp$p)
}
\arguments{
\item{bp}{Object}

\item{which.var}{which variable(s) to find coordinates}
}
\value{
Axes coordinates
}
\description{
Provide axes coordinates
}
